(feat:world) Finish reading Anvil worlds

+++ src/main/java/moe/seikimo/droplet/utils/objects/binary/BitArray.java
@@ -43,7 +43,7 @@ public interface BitArray {
         VarInts.writeInt(buffer, size);
     }
 
-    int[] getWords();
+    long[] getWords();
 
     BitArrayVersion getVersion();
